move runWithoutPipes() to comby_test.go
